+++ zerver/actions/presence.py
@@ -1,6 +1,5 @@
 import datetime
 import time
-from typing import Optional
 
 from django.conf import settings
 from django.db import transaction
@@ -9,8 +8,7 @@ from zerver.actions.user_activity import update_user_activity_interval
 from zerver.decorator import statsd_increment
 from zerver.lib.queue import queue_json_publish
 from zerver.lib.timestamp import datetime_to_timestamp
-from zerver.lib.user_status import update_user_status
-from zerver.models import Client, UserPresence, UserProfile, UserStatus, active_user_ids, get_client
+from zerver.models import Client, UserPresence, UserProfile, active_user_ids, get_client
 from zerver.tornado.django_api import send_event
 
 
@@ -146,49 +144,3 @@ def update_user_presence(
 
     if new_user_input:
         update_user_activity_interval(user_profile, log_time)
-
-
-def do_update_user_status(
-    user_profile: UserProfile,
-    away: Optional[bool],
-    status_text: Optional[str],
-    client_id: int,
-    emoji_name: Optional[str],
-    emoji_code: Optional[str],
-    reaction_type: Optional[str],
-) -> None:
-    if away is None:
-        status = None
-    elif away:
-        status = UserStatus.AWAY
-    else:
-        status = UserStatus.NORMAL
-
-    realm = user_profile.realm
-
-    update_user_status(
-        user_profile_id=user_profile.id,
-        status=status,
-        status_text=status_text,
-        client_id=client_id,
-        emoji_name=emoji_name,
-        emoji_code=emoji_code,
-        reaction_type=reaction_type,
-    )
-
-    event = dict(
-        type="user_status",
-        user_id=user_profile.id,
-    )
-
-    if away is not None:
-        event["away"] = away
-
-    if status_text is not None:
-        event["status_text"] = status_text
-
-    if emoji_name is not None:
-        event["emoji_name"] = emoji_name
-        event["emoji_code"] = emoji_code
-        event["reaction_type"] = reaction_type
-    send_event(realm, event, active_user_ids(realm.id))
